FIX) 로그아웃 API 오류 수정개발

+++ src/main/java/kr/imcf/ott/account/oauth2/logout/LogoutServiceImpl.java
@@ -8,6 +8,7 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 
+import javax.swing.text.html.Option;
 import java.util.Optional;
 
 @Slf4j
@@ -21,16 +22,26 @@ public class LogoutServiceImpl implements LogoutService {
     @Override
     public boolean logout(String token) {
         OAuth2Principal oAuth2Principal = null;
+        Optional<RAccount> currentTokenInfo = Optional.empty();
 
         try {
             oAuth2Principal = jwtProvider.toPrincipal(token);
-            Optional<RAccount> currentTokenInfo = accountRedisRepository.findById(RAccount.idFormat(oAuth2Principal.getPlatformType(), oAuth2Principal.getEmail()));
-            accountRedisRepository.delete(currentTokenInfo.get());
+            currentTokenInfo = accountRedisRepository.findById(RAccount.idFormat(oAuth2Principal.getPlatformType(), oAuth2Principal.getEmail()));
+
+            if(currentTokenInfo.isPresent() && currentTokenInfo.get().getAccessToken().equals(token)){
+                accountRedisRepository.delete(currentTokenInfo.get());
+                log.info(String.format("Logout Complete :: %s", token));
+                return true;
+            }
+            else{
+                log.error(String.format("Token is not existed :: %s", token));
+            }
+
         } catch (Exception e) {
-            log.info(String.format("Parsing Failed :: %s", token));
-            return false;
+            log.error(String.format("Parsing Failed :: %s", token));
         }
-        return true;
+
+        return false;
     }
 
 }
